# docs/contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {"title": "跃科网络空间安全实训平台 API", "version": "1.0.0"},
  "paths": {
...
    "/api/v1/resources/files": {"post": {"summary": "上传课程资源真实文件并登记文件对象", "responses": {"201": {"description": "文件对象"}, "413": {"description": "文件超过限制"}, "422": {"description": "文件为空或类型不受支持"}}}},
...
    "/api/v1/resources/{resource_id}/download": {"get": {"summary": "下载受控存储中的资源版本文件", "responses": {"200": {"description": "真实文件"}, "404": {"description": "文件不存在或无权访问"}}}},
...
  },
...
}